Moved Integer prims from interp.c to prim.c
Moving primitive operations over to prim.c slowly.
