+++ src/vm/interp.c
@@ -207,72 +207,6 @@ static struct object *lookupMethod(struct object *selector, struct object *class
 
 
 /*
- * newLInteger()
- *  Create new Integer (64-bit)
- */
-static struct object *newLInteger(int64_t val)
-{
-    struct object *res;
-    int64_t *tmp;
-
-    res = gcialloc(sizeof(int64_t));
-    res->class = IntegerClass;
-    tmp = (int64_t *)bytePtr(res);
-    *tmp = val;
-    return(res);
-}
-
-/*
- * do_Integer()
- *  Implement the appropriate 64-bit Integer operation
- *
- * Returns NULL on error, otherwise the resulting Integer or
- * Boolean (for comparisons) object.
- */
-static struct object *do_Integer(int op, struct object *low, struct object *high)
-{
-    int64_t l, h;
-    int64_t *tmp;
-
-    tmp = (int64_t *)bytePtr(low);
-    l = *tmp;
-    tmp = (int64_t *)bytePtr(high);
-    h = *tmp;
-    switch (op) {
-    case 25:    /* Integer division */
-        if (h == 0LL) {
-            return(NULL);
-        }
-        return(newLInteger(l/h));
-
-    case 26:    /* Integer remainder */
-        if (h == 0LL) {
-            return(NULL);
-        }
-        return(newLInteger(l%h));
-
-    case 27:    /* Integer addition */
-        return(newLInteger(l+h));
-
-    case 28:    /* Integer multiplication */
-        return(newLInteger(l*h));
-
-    case 29:    /* Integer subtraction */
-        return(newLInteger(l-h));
-
-    case 30:    /* Integer less than */
-        return((l < h) ? trueObject : falseObject);
-
-    case 31:    /* Integer equality */
-        return((l == h) ? trueObject : falseObject);
-
-    default:
-        error("do_Integer(): Invalid primitive integer operation %d!", op);
-    }
-    return(NULL);
-}
-
-/*
  * bulkReplace()
  *  Implement replaceFrom:to:with:startingAt: as a primitive
  *
